Generalized LSR alg. and took of self-tuning.

+++ SSC_Arduino/RLSR.h
@@ -1,6 +1,6 @@
 typedef struct RLSR
 {
-  unsigned long n;
+  unsigned int n;
   double D1;
   double D2;
   double sx;
